feat: AdminPage SRP 적용 및 컴포넌트 계층 구조 완성
- AdminPage: 라우팅 + 탭 관리로 역할 단순화
- 엔티티 컴포넌트: ProductManagement, CouponManagement
- UI 컴포넌트: AdminTabs
- 폼 컴포넌트: ProductForm, CouponForm
- localStorage: selectedCoupon 상태 유지
